requirements.txt
geopandas==0.9.0
haversine==2.3.0
matplotlib==3.4.1
networkx==2.5.1
osmnx==1.1.1
pandas==1.2.4
pickleshare==0.7.5
python-csv==0.0.13
python-telegram-bot==13.4.1
scikit-learn==0.24.2
staticmap==0.5.5
urllib3==1.25.8
